* Merged MIXER CROP, MIXER ANCHOR, MIXER ROTATION and MIXER PERSPECTIVE from 2.0 * Merged reporting of current MIXER values when no parameters are given. * Changed so that each vertex / texture coordinate pair in frame_geometry is encapsulated in a struct, to not require producers to keep track of indexes in an a array. * Changed so that the same code in image_kernel draws a single QUAD and a list of QUAD's * TODO make perspective correction and cropping work with non default geometry like the text producer. Currently no cropping or perspective correction works for custom geometry. * TODO make interaction coordinate translations work with new transforms. * Reverted text_producer change so that the baseline of the font is the actual position. * Fixed so that JSON encoded template data is also forwarded as is, instead of being interpreted as an .ftd file name. * Implemented anchor, crop, perspective and rotation support in scene_producer

+++ core/producer/text/utils/texture_atlas.cpp
@@ -138,7 +138,7 @@ public:
 	}
 
 	//the data parameter points to bitmap-data that is 8-bit grayscale.
-	void set_region(const size_t x, const size_t y, const size_t width, const size_t height, const unsigned char *src, const size_t stride, const color<float>& col)
+	void set_region(const size_t x, const size_t y, const size_t width, const size_t height, const unsigned char *src, const size_t stride, const color<double>& col)
 	{
 		//this assumes depth_ is set to 4
 		for(size_t i=0; i<height; ++i)
@@ -216,7 +216,7 @@ private:
 
 texture_atlas::texture_atlas(const size_t w, const size_t h, const size_t d) : impl_(new impl(w, h, d)) {}
 rect texture_atlas::get_region(int width, int height) { return impl_->get_region(width, height); }
-void texture_atlas::set_region(const size_t x, const size_t y, const size_t width, const size_t height, const unsigned char *src, const size_t stride, const color<float>& col)
+void texture_atlas::set_region(const size_t x, const size_t y, const size_t width, const size_t height, const unsigned char *src, const size_t stride, const color<double>& col)
 	{ impl_->set_region(x, y, width, height, src, stride, col); }
 
 void texture_atlas::clear() { impl_->clear(); }
